+++ inc/strongship.h
@@ -16,8 +16,8 @@ public:
     void DecreaseFuel();
     double GetHealth() override;
     int GetFuel() const override;
-    int GetChance() const override;
-    int GetDamage() const override;
+    double GetChance() const override;
+    double GetDamage() const override;
     int GetMoney() override;
     void changeMoney(int pay) override;
 private:
